requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
oracledb==2.2.1
mysql-connector-python==9.0.0
pyodbc==5.1.0
psycopg2-binary==2.9.9
